+++ framework_src/native_library/android/FreKotlinExampleANE/FreKotlin/src/main/java/com/adobe/air/FreKotlinActivityResultCallback.java
@@ -15,6 +15,6 @@
  */
 
 package com.adobe.air;
-public interface TRActivityResultCallback extends AndroidActivityWrapper.ActivityResultCallback {
+public interface FreKotlinActivityResultCallback extends AndroidActivityWrapper.ActivityResultCallback {
 
 }