+++ framework_src/native_library/android/FreKotlinExampleANE/FreKotlin/src/main/java/com/adobe/air/FreKotlinStateChangeCallback.java
@@ -15,5 +15,5 @@
  */
 
 package com.adobe.air;
-public interface TRStateChangeCallback extends AndroidActivityWrapper.StateChangeCallback {
+public interface FreKotlinStateChangeCallback extends AndroidActivityWrapper.StateChangeCallback {
 }