+++ starter_project/native_library/android/HelloWorldANE/app/src/main/java/com/mycompany/helloworldane/KotlinController.kt
@@ -1,17 +1,23 @@
 package com.mycompany.helloworldane
 
+import android.content.Intent
+import android.content.res.Configuration
+import com.adobe.air.AndroidActivityWrapper.ActivityState
+import com.adobe.air.AndroidActivityWrapper.ActivityState.*
+import com.adobe.air.FreKotlinActivityResultCallback
+import com.adobe.air.FreKotlinStateChangeCallback
 import com.adobe.fre.FREContext
 import com.adobe.fre.FREObject
 import com.tuarua.frekotlin.*
 
 @Suppress("unused", "UNUSED_PARAMETER", "UNCHECKED_CAST")
-class KotlinController : FreKotlinMainController {
+class KotlinController : FreKotlinMainController, FreKotlinStateChangeCallback, FreKotlinActivityResultCallback {
     fun init(ctx: FREContext, argv: FREArgv): FREObject? {
         return null
     }
 
     fun sayHello(ctx: FREContext, argv: FREArgv): FREObject? {
-        argv.takeIf { argv.size > 2 } ?: return FreArgException("sayHello")
+        argv.takeIf { argv.size > 2 } ?: return FreArgException()
 
         val myString = String(argv[0])
         val uppercase = Boolean(argv[1])
@@ -34,31 +40,24 @@ class KotlinController : FreKotlinMainController {
         return null
     }
 
-    override fun onStarted() {
-        super.onStarted()
+    override fun onActivityResult(requestCode: Int, resultCode: Int, intent: Intent?) {
     }
 
-    override fun onRestarted() {
-        super.onRestarted()
+    override fun onConfigurationChanged(configuration: Configuration?) {
     }
 
-    override fun onResumed() {
-        super.onResumed()
+    override fun onActivityStateChanged(activityState: ActivityState?) {
+        when (activityState) {
+            STARTED -> return
+            RESTARTED -> return
+            RESUMED -> return
+            PAUSED -> return
+            STOPPED -> return
+            DESTROYED -> return
+        }
     }
 
-    override fun onPaused() {
-        super.onPaused()
-    }
-
-    override fun onStopped() {
-        super.onStopped()
-    }
-
-    override fun onDestroyed() {
-        super.onDestroyed()
-    }
-
-    override val TAG: String
+    override val TAG: String?
         get() = this::class.java.simpleName
     private var _context: FREContext? = null
     override var context: FREContext?